Always attach volumes to a major device.
Attaching volumes to minor devices (e.g. /dev/sdf1) causes problems if
the volume has a partition table. We can always attach to the major
device (/dev/sdf) then use blkid to check for a filesystem. If it doesn't
find one on the major device, assume the filesystem we want is on the
first partition/minor device (e.g. /dev/sdf1).